Update main.py
Added basic YouTube video functionality. Next up: converting everything to embeds, (maybe) adding the functionality to react to a message to get lyrics so that it doesn't kill chat. Adding a queue command so that it doesn't get clogged, and also maybe an add/remove command to add/remove songs from the queue.