Update main.py
Updated queue functionality, the queue is now fully operational and goes to the next song after one is done playing. Songs with lyrics less than 6,000 characters now have fully informational embeds, including artist name and release date. Next up: Multi-page embeds for songs with 6,000 character plus lyrics, and a skip function to skip through songs in the queue. Also, something to think about is editing the queue's embed to include the names of the videos. That will be done through editing what is added to the queue before it is added.